cleans up user/kb/quantum calls, keyboard files
