+++ keyboards/alps64/alps64.c
@@ -20,15 +20,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define LED_OFF()   do { DDRC &= ~(1<<5); PORTC &= ~(1<<5); } while (0)
 #define LED_TGL()   do { DDRC |= (1<<5); PINC |= (1<<5); } while (0)
 
-__attribute__ ((weak))
-void matrix_init_user(void) {
-
-}
-
-__attribute__ ((weak))
-void matrix_scan_user(void) {
-
-}
 
 void matrix_init_kb(void) {
     LED_ON();
@@ -36,8 +27,4 @@ void matrix_init_kb(void) {
     LED_OFF();
     
     matrix_init_user();
-}
-
-void matrix_scan_kb(void) {
-    matrix_scan_user();
 }
