+++ keyboards/atomic/config.h
@@ -46,6 +46,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define MATRIX_COL_PINS { F1, F0, B0, C7, F4, F5, F6, F7, D4, D6, B4, D7, D3, D2, D1 }
 #define UNUSED_PINS
 
+#define BACKLIGHT_PIN B7
+
 /* COL2ROW or ROW2COL */
 #define DIODE_DIRECTION COL2ROW
