+++ keyboards/ergodox_ez/ergodox_ez.c
@@ -4,16 +4,6 @@
 bool i2c_initialized = 0;
 uint8_t mcp23018_status = 0x20;
 
-__attribute__ ((weak))
-void matrix_init_user(void) {
-
-}
-
-__attribute__ ((weak))
-void matrix_scan_user(void) {
-
-}
-
 void matrix_init_kb(void) {
    // keyboard LEDs (see "PWM on ports OC1(A|B|C)" in "teensy-2-0.md")
     TCCR1A = 0b10101001;  // set and configure fast PWM
@@ -36,11 +26,6 @@ void matrix_init_kb(void) {
 
     matrix_init_user();
 }
-
-void matrix_scan_kb(void) {
-    matrix_scan_user();
-}
-
 
 void ergodox_blink_all_leds(void)
 {
