+++ keyboards/ergodox_ez/ergodox_ez.h
@@ -119,9 +119,4 @@ inline void ergodox_led_all_set(uint8_t n)
     { k0D, k1D, k2D, k3D, k4D, KC_NO }    \
    }
 
-void matrix_init_user(void);
-void matrix_scan_user(void);
-
-
-
 #endif
