+++ keyboards/jd45/config.h
@@ -37,6 +37,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define MATRIX_COL_PINS { F4, D7, B5, B6, C6, C7, D4, D6, D5, D0, D1, D2, B0 }
 #define UNUSED_PINS
 
+#define BACKLIGHT_PIN B7
+
 /* COL2ROW or ROW2COL */
 #define DIODE_DIRECTION COL2ROW
