+++ keyboards/kc60/config.h
@@ -50,6 +50,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define MATRIX_COL_PINS { F0, F1, E6, C7, C6, B7, D4, B1, B0, B5, B4, D7, D6, B3 }
 #define UNUSED_PINS
 
+#define BACKLIGHT_PIN B6
+
 /* COL2ROW or ROW2COL */
 #define DIODE_DIRECTION COL2ROW
