+++ keyboards/preonic/preonic.h
@@ -33,9 +33,4 @@
 	{ k40, k41, k42, k43, k44, k45, k46, k47, k48, k49, k4a, k4b } \
 }
 
-void matrix_init_user(void);
-void matrix_scan_user(void);
-bool process_action_kb(keyrecord_t *record);
-void backlight_init_ports(void);
-
 #endif
