+++ keyboards/retro_refit/retro_refit.c
@@ -1,19 +1,5 @@
 #include "retro_refit.h"
-
-__attribute__ ((weak))
-void matrix_init_user(void) {
-    // leave this function blank - it can be defined in a keymap file
-};
-
-__attribute__ ((weak))
-void matrix_scan_user(void) {
-    // leave this function blank - it can be defined in a keymap file
-};
-
-__attribute__ ((weak))
-void led_set_user(uint8_t usb_led) {
-    // leave this function blank - it can be defined in a keymap file
-};
+#include "led.h"
 
 void matrix_init_kb(void) {
     // put your keyboard start-up code here
@@ -24,13 +10,6 @@ void matrix_init_kb(void) {
     PORTD |= (1<<6);
 
     matrix_init_user();
-};
-
-void amatrix_scan_kb(void) {
-    // put your looping keyboard code here
-    // runs every cycle (a lot)
-
-    matrix_scan_user();
 };
 
 void led_set_kb(uint8_t usb_led) {
