+++ keyboards/sixkeyboard/matrix.c
@@ -40,10 +40,20 @@ static matrix_row_t matrix[MATRIX_ROWS];
 
 __attribute__ ((weak))
 void matrix_init_kb(void) {
+    matrix_init_user();
 }
 
 __attribute__ ((weak))
 void matrix_scan_kb(void) {
+    matrix_scan_user();
+}
+
+__attribute__ ((weak))
+void matrix_init_user(void) {
+}
+
+__attribute__ ((weak))
+void matrix_scan_user(void) {
 }
 
 inline
