+++ keyboards/sixkeyboard/sixkeyboard.h
@@ -15,8 +15,4 @@
 // instead of the obtuse 11x8 matrix.
 
 
-void matrix_init_user(void);
-void matrix_scan_user(void);
-void led_set_user(uint8_t usb_led);
-
 #endif
